Partial fix for SSGD-2 (see description)
- The SSGD-2 ticket has a comment about what the problem is.
- This check-in is only to correct for the dialog showing.  A world at
the end of the world list that doesn't have any levels will no longer
appear to have a bunch of them.
- With this check-in, if you create a world without any levels, go to an
editor (i.e. Projectile Editor) and then come back, a level will be
added to the world.  You will not be able to delete it.  This will be
corrected for in another check-in.