SSGD-67 Fix for adding phantom level
The old code I deleted was forcing the world index to 1 if there are
more than 2 worlds built for the game.  Having the Unused Levels world
and two regular worlds met this condition.  But now we're using a new
world index but keeping the old level index.  This causes the level
container to select the Add Levels button, thus adding a level to World
1.

I don't know why we would want to select the first valid world when
there are multiple worlds, so removed this code.  Mixing world indices
and level indices is just bad.